# By leanid (11) and others
# Via leanid (3) and Leanid Chaika (1)
* commit '66910fceff511daf076a03d64a7b56e6ec5c3656': (21 commits)
  DF-6084 disable warnings as errors temporary continue in other task
  DF-6084 fix one more warnig
  DF-6084 fix ios build with cmake
  DF-6084 fix cmake flags
  DF-6084 start adding treat warnings as errors
  DF-6084 continue fix cmake treat warnings as error
  DF-6084 enable warnings as errors in all unit test with cmake
  DF-6084 fix review comments
  DF-6084 fix build resource editor
  DF-6084 fix win32 unit test
  DF-6084 fix compile, next need fix link
  DF-6084 continue
  DF-6084 continue fix warnings ios
  DF-6084 continue fix warnings
  DF-6084 resolve Save virtual name
  DF-6084 rename to resolve conflict MongoDB SaveObject and BaseObject SaveObject
  DF-6084 rename Load/Save BaseObject methods to fix warnings
  DF-6084 continue fix unittests warnings
  DF-6084 fix review comments. set warnings as errors on ios, android, win32
  DF-6084 fix review comments
  ...

+++ Sources/Internal/Base/IntrospectionCollection.h
@@ -63,7 +63,7 @@ namespace DAVA
 
 			if(NULL != object)
 			{
-				size = ((CollectionT *) object)->size();
+				size = static_cast<int>(((CollectionT *) object)->size());
 			}
 
 			return size;
